Add ability to send metadata from issues (#37)
The current planned procedure, consisting in sending json files via pull requests, has security issues as shown in PR #35
This adds an issue template and GitHub workflows to create pull requests and enable auto merge from issues, checking that publisher is not modified and it's created based on the GitHub account of the issue creator. This publisher can edit metadata send by the corresponding person, for example to change the add-on channel from beta to stable once the add-on has been tested.