Add support for CodeQL (#2660)
fixes #2655

Summary of the issue
Some vulnerabilities can be detected analyzing source code of add-ons submitted to the store.

Development strategy
After validating add-on metadata, Upload artifact action is used to upload the add-on previously downloaded to be validated.
A workflow to analyze the source code with CodeQL is called.
This workflow calls the Download artifact action to download the add-on.
The addon.nvda-addon file is moved to an addon.zip file to be expanded for the analysis.
The source code is analyzed. Languages supported in the used workflow are just Python and JavaScript, since they are interpreted languages which can be used in add-ons. Other languages like C++ may require .dll files, so maybe faster just to analyze the mentioned languages.
If analysis fails, a comment is posted to the issue requesting the submitter to contact NV Access for more details (failures aren't tested now).
If the job succeeds, the workflow to check metadata and create a pull request with the .json file is called.