Add steps outlining the review process for NV Access for new add-on submissions (#3369)
Documentation is missing for the process NV Access follows when approving authors to submit to an add-on ID

When a GitHub user submits a particular add-on for the first time, the submission is blocked pending initial review from NV Access.
This is to confirm that the GitHub user has authorisation to submit the add-on from the add-on maintainers.
In some cases, 2 separate add-on maintainers may want to lay claim to a single "add-on ID".
It is important for users to expect consistent authorship for a single add-on ID for trust and security reasons.
It would be misleading and potentially risky to be encouraged to update to a fork or alternative add-on without permission from the original maintainers.
See the submission guide for further reasoning behind this.
This registration process ensures a specific add-on ID is only submitted by the same group of authorised maintainers.