Raise error if no files are found (#50)
Passing in an invalid path leads validation to unexpectedly pass.
This is because when globbing the path, no files are found, meaning validation occurs against no files.
This can be quite misleading for developers, as it is unclear why passing in what seems like a proper file path with validation errors yields no validation errors. See nvaccess/addon-datastore#6712 for context